+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certifiedProduct/validation/InpatientModular2014Validator.java
@@ -57,7 +57,15 @@ public class InpatientModular2014Validator extends CertifiedProduct2014Validator
 		
 		for(PendingCertificationResultDTO cert : product.getCertificationCriterion()) {
 			if(cert.getMeetsCriteria() != null && cert.getMeetsCriteria() == Boolean.TRUE) {
-				if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_TOOLS_USED) &&
+				
+				boolean gapEligibleAndTrue = false;
+				if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.GAP) &&
+						cert.getGap() == Boolean.TRUE) {
+					gapEligibleAndTrue = true;
+				}
+				
+				if(!gapEligibleAndTrue && 
+						certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_TOOLS_USED) &&
 						!cert.getNumber().equals("170.314 (g)(1)") && 
 						!cert.getNumber().equals("170.314 (g)(2)") &&
 						(cert.getTestTools() == null || cert.getTestTools().size() == 0)) {
@@ -119,6 +127,30 @@ public class InpatientModular2014Validator extends CertifiedProduct2014Validator
 		
 		if(hasG1Cert && hasG2Cert) {
 			product.getWarningMessages().add("Both (g)(1) and (g)(2) were found which is not typically permitted.");
+		}
+	}
+	
+	@Override
+	protected void validateDemographics(CertifiedProductSearchDetails product) {
+		super.validateDemographics(product);
+		
+		for(CertificationResult cert : product.getCertificationResults()) {
+			if(cert.isSuccess() != null && cert.isSuccess() == Boolean.TRUE) {
+				
+				boolean gapEligibleAndTrue = false;
+				if(certRules.hasCertOption(cert.getNumber(), CertificationResultRules.GAP) &&
+						cert.isGap() == Boolean.TRUE) {
+					gapEligibleAndTrue = true;
+				}
+				
+				if(!gapEligibleAndTrue && 
+						certRules.hasCertOption(cert.getNumber(), CertificationResultRules.TEST_TOOLS_USED) &&
+						!cert.getNumber().equals("170.314 (g)(1)") && 
+						!cert.getNumber().equals("170.314 (g)(2)") &&
+						(cert.getTestToolsUsed() == null || cert.getTestToolsUsed().size() == 0)) {
+						product.getErrorMessages().add("Test Tools are required for certification " + cert.getNumber() + ".");
+				}
+			}
 		}
 	}